[MOVE] types 폴더 루트로 이동
