+++ src/auth/service/common-auth.provider.ts
@@ -1,5 +1,5 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { GoogleIdTokenPayload, JWT } from '../types/auth';
+import { GoogleIdTokenPayload, JWT } from '../../types/auth';
 
 @Injectable()
 export class CommonAuthService {
