+++ src/user/user.controller.ts
@@ -27,7 +27,7 @@ import {
 } from './dto/response/response.dto';
 import { KakaoSignOutReqDto } from '@src/user/dto/request/kakao-req.dto';
 import { NicknameReqDto, UidReqDto } from './dto/request/user-req.dto';
-import { AuthToken } from './types/user';
+import { AuthToken } from '../types/user';
 import { AuthGuard } from '@src/guards/auth.guard';
 
 @ApiTags('user')
